backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.4.4
